refactor(#251): features/team refactor
