+++ src/widgets/home/FavoriteGoalsSection/FavoriteGoalsItem.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { GoalProgressCard } from "@/widgets/team/GoalProgressCard/GoalProgressCard";
+import { GoalProgressCard } from "@/widgets/team/GoalProgressCard";
 
 export interface FavoriteGoalsItemProps {
   teamId: number;
